src/serveur/social/pyproject.toml
[project]
name = "demperm-social-backend"
version = "1.0.0"
description = "Backend API pour réseau social politique local"
authors = [
    {name = "EPITA EGEO Team", email = "team@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.14,<4.0",
    "djangorestframework-simplejwt>=5.3,<6.0",
    "django-cors-headers>=4.3,<5.0",
    "django-ratelimit>=4.1,<5.0",
    "psycopg[binary]>=3.1,<4.0",
    "redis>=5.0,<6.0",
    "cryptography>=41.0,<42.0",
    "bleach>=6.1,<7.0",
    "python-jose[cryptography]>=3.3,<4.0",
    "requests>=2.31,<3.0",
    "requests-oauthlib>=1.3,<2.0",
    "python-dotenv>=1.0,<2.0",
    "gunicorn>=21.2,<22.0",
    "drf-yasg>=1.21,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<8.0",
    "pytest-django>=4.7,<5.0",
    "pytest-cov>=4.1,<5.0",
    "pytest-asyncio>=0.21,<1.0",
    "black>=23.12,<24.0",
    "flake8>=6.1,<7.0",
    "mypy>=1.7,<2.0",
    "django-stubs>=4.2,<5.0",
    "types-requests>=2.31,<3.0",
    "ipython>=8.18,<9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = [
    "migrations/",
    "tests/",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "django.settings"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["api/tests"]
addopts = [
    "--verbose",
    "--strict-markers",
    "--cov=api",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["api"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/__init__.py",
    "*/manage.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

